# internal/router/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/burnFile": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件"
                ],
                "summary": "下载文件",
                "parameters": [
                    {
                        "description": " ",
                        "name": "fileReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BurnFileDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/burnMixVideo": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "三合一"
                ],
                "summary": "下载三合一画面",
                "parameters": [
                    {
                        "description": " ",
                        "name": "mix3Req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BurnMix3VideoDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/burnOtherSingleParams": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "下载参数"
                ],
                "summary": "公区",
                "parameters": [
                    {
                        "description": " ",
                        "name": "publicReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BurnSingleVideoDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/burnOtherSingleVideo": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "公区"
                ],
                "summary": "下载公区",
                "parameters": [
                    {
                        "description": " ",
                        "name": "publicReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BurnSingleVideoDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/burnParams": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "下载参数"
                ],
                "summary": "三合一",
                "parameters": [
                    {
                        "description": " ",
                        "name": "burnMix3VideoDTO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BurnMix3VideoDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/burnSingleParams": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "下载参数"
                ],
                "summary": "房间单画面",
                "parameters": [
                    {
                        "description": " ",
                        "name": "singleReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BurnSingleVideoDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/burnSingleVideo": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "房间单画面"
                ],
                "summary": "下载房间单画面",
                "parameters": [
                    {
                        "description": " ",
                        "name": "singleReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BurnSingleVideoDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/burnTask": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "刻录任务"
                ],
                "summary": "创建刻录任务",
                "parameters": [
                    {
                        "description": " ",
                        "name": "burnSettingDTO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BurnSettingDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.BurnFileDTO": {
            "type": "object",
            "properties": {
                "callbackStatus": {
                    "type": "integer"
                },
                "callbackUrl": {
                    "type": "string",
                    "example": "http://localhost:8010/api/v1/callback"
                },
                "done_status": {
                    "type": "integer"
                },
                "fileUrl": {
                    "type": "string",
                    "example": "https://static.runoob.com/images/demo/demo1.jpg"
                },
                "odaSavePath": {
                    "type": "string",
                    "example": "D:/downloadVideo"
                },
                "taskId": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "dto.BurnMix3VideoDTO": {
            "type": "object",
            "required": [
                "endTime",
                "rtspUrlLeft",
                "rtspUrlMiddle",
                "rtspUrlRight",
                "startTime"
            ],
            "properties": {
                "callBackUrl": {
                    "type": "string",
                    "example": "http://localhost:8010/api/v1/callback"
                },
                "endTime": {
                    "type": "string",
                    "example": "2022-05-11 15:25:00"
                },
                "fileSavePath": {
                    "type": "string"
                },
                "odaSavePath": {
                    "type": "string",
                    "example": "D:/downloadVideo"
                },
                "rtspUrlLeft": {
                    "type": "string",
                    "example": "rtsp://admin:cebon61332433@192.168.99.215:554/cam/realmonitor?channel=1\u0026subtype=1"
                },
                "rtspUrlMiddle": {
                    "type": "string",
                    "example": "rtsp://admin:cebon61332433@192.168.99.215:554/cam/realmonitor?channel=1\u0026subtype=0"
                },
                "rtspUrlRight": {
                    "type": "string",
                    "example": "rtsp://admin:cebon61332433@192.168.99.215:554/cam/realmonitor?channel=1\u0026subtype=1"
                },
                "startTime": {
                    "type": "string",
                    "example": "2022-05-11 15:20:00"
                },
                "taskId": {
                    "type": "integer",
                    "example": 1
                },
                "temperature": {
                    "type": "string",
                    "example": ""
                },
                "videoName": {
                    "type": "string"
                }
            }
        },
        "dto.BurnSettingDTO": {
            "type": "object",
            "required": [
                "odaSavePath",
                "taskNum"
            ],
            "properties": {
                "encryptionType": {
                    "type": "integer"
                },
                "odaSavePath": {
                    "type": "string",
                    "example": "D:/videodata"
                },
                "password": {
                    "type": "string"
                },
                "taskNum": {
                    "type": "integer",
                    "example": 1
                },
                "videoName": {
                    "type": "string",
                    "example": "video"
                }
            }
        },
        "dto.BurnSingleVideoDTO": {
            "type": "object",
            "required": [
                "endTime",
                "rtspUrl",
                "startTime"
            ],
            "properties": {
                "callBackUrl": {
                    "type": "string",
                    "example": "http://localhost:8010/api/v1/callback"
                },
                "endTime": {
                    "type": "string",
                    "example": "2022-05-11 15:25:00"
                },
                "fileSavePath": {
                    "type": "string"
                },
                "odaSavePath": {
                    "type": "string",
                    "example": "D:/downloadVideo"
                },
                "rtspUrl": {
                    "type": "string",
                    "example": "rtsp://admin:cebon61332433@192.168.99.215:554/cam/realmonitor?channel=1\u0026subtype=0"
                },
                "startTime": {
                    "type": "string",
                    "example": "2022-05-11 15:20:00"
                },
                "taskId": {
                    "type": "integer",
                    "example": 1
                },
                "videoName": {
                    "type": "string"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "msg": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        }
    }
}